requirements.txt
flask
flask_sqlalchemy
sqlalchemy
pymysql
opencv-python
opencv_contrib_python
pillow
tensorflow==1.15.0
keras==2.3.1
scipy==1.2.2
scikit-learn==0.21.3
sklearn
mtcnn
keyboard
h5py==2.10.0
imageio
imutils
matplotlib
pandas
numpy==1.19.3
joblib==0.17.0
playsound
pyttsx3
six
seaborn
setuptools
wheel
dlib